+++ nuxt.config.ts
@@ -7,6 +7,7 @@ const normalizedBase = appBaseURL.endsWith('/') ? appBaseURL : `${appBaseURL}/`
 export default defineNuxtConfig({
   compatibilityDate: '2025-07-15',
   future: { compatibilityVersion: 4 },
+  modules: ['@vueuse/nuxt'],
   devtools: { enabled: true },
   css: ['~/assets/css/main.css'],
   app: {